+++ src/components/answers/AnswerPresentationPage.jsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useNavigate, useParams } from "react-router";
-import { Button, TextField, Typography } from "@mui/material";
+import { Button, Typography } from "@mui/material";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import PresentationService from "../../service/PresentationService";
 import { MCQAnswerSection } from "./MCQ/MCQAnswerSection";
 import { slideTypes } from "../EditSlidePanel/EditSlidePanel";
-import AnswerService from "../../service/AnswerService";
+import { WordCloudAnswersSection } from "./word-cloud/WordCloudAnswersSection";
 
 function SuccessScreen() {
   return (
@@ -23,45 +23,6 @@ function SuccessScreen() {
       <Typography style={{ position: "absolute", bottom: 0 }} variant="body2">
         ¿La diapositiva no ha cambiado?<Button>Recarga la pagina</Button>
       </Typography>
-    </>
-  );
-}
-
-function WordCloudAnswersSection({
-  question,
-  slideId,
-  setSuccess,
-  startPolling,
-}) {
-  const [text, setText] = useState("");
-
-  function handleSubmit() {
-    AnswerService.answer(text, slideId).then(() => {
-      setSuccess(true);
-      startPolling();
-    });
-  }
-
-  return (
-    <>
-      <Typography style={{ marginBottom: "1em" }} variant="h5">
-        {question || ""}
-      </Typography>
-      <div>
-        <TextField
-          placeholder="Escribí aca tu respuesta"
-          sx={{ width: "40vw", marginBottom: "2em" }}
-          value={text}
-          onChange={(e) => setText(e.target.value)}
-        />
-      </div>
-      <Button
-        variant="contained"
-        sx={{ width: "24em" }}
-        onClick={() => handleSubmit()}
-      >
-        Enviar
-      </Button>
     </>
   );
 }
